goodbye old magicgui
