+++ examples/basic.py
@@ -1,12 +1,4 @@
-from qtpy.QtWidgets import QApplication
-
-from magicgui import magicgui
-
-running = True
-app = QApplication.instance()
-if not app:
-    running = False
-    app = QApplication([])
+from magicgui import event_loop, magicgui
 
 
 @magicgui
@@ -14,7 +6,5 @@ def example(arg=1):
     return arg
 
 
-widget = example.Gui(show=True)
-
-if not running:
-    app.exec_()
+with event_loop():
+    example.show()
