+++ examples/basic2.py
@@ -16,20 +16,12 @@ class Medium(Enum):
     Air = 1.0003
 
 
-def choices(obj):
-    return ["a", "b", "c"]
-
-
-@magicgui(
-    call_button="calculate",
-    aoi={"widget_type": "FloatSlider", "maximum": 10},
-    b={"choices": choices},
-)
+@magicgui(call_button="calculate")
 def snells_law(
-    aoi=30.0, n1=Medium.Glass, n2=Medium.Water, b: str = "c", degrees=True,
+    aoi=1.0, n1=Medium.Glass, n2=Medium.Water, degrees=True,
 ):
     """Calculate the angle of refraction given two media and an AOI."""
-    print(aoi)
+    print("aoi in function", aoi)
     if degrees:
         aoi = math.radians(aoi)
     try:
