+++ magicgui/_coreOLD.py
