--- magicgui/_qt/types.py
@@ -1,6 +1,0 @@
-"""Types for export."""
-from qtpy.QtCore import Signal
-from qtpy.QtWidgets import QPushButton
-
-ButtonType = QPushButton
-SignalType = Signal
