+++ magicgui/_qtOLD/eval_lineedit.py
