+++ setup.cfg
@@ -28,7 +28,7 @@ implicit_reexport = False
 # disallow_untyped_defs = True
 show_column_numbers = True
 show_error_codes = True
-pretty = True
+# pretty = True
 
 [mypy-magicgui._tests.*]
 ignore_errors = True
@@ -44,5 +44,4 @@ ignore_missing_imports = True
 
 [isort]
 profile = black
-known_third_party = napari,numpy,numpydoc,pytest,qtpy,setuptools,skimage,typing_extensions,vispy,wrapt
 src_paths=magicgui
